Start implementing CHKN practice tool
This will help the CHKN beginner get started typing. This commit only has a
boilerplate web page which shows an outline of a keyboard. The keyboard will
show the Kanji that can be typed with one more letter after the currently
pending (uncommitted) keys.

This tool will be able to run without a webserver: just open the HTML page
locally in your browser to get started.